Turn off CSS and JS preprocessing
CSS and JS caching are resulting in 404s (and styling, etc. not
looking and behaving as expected). This commit turns off CSS and
JS preprocessing in order to have theme assets load as intended.

Once https://github.com/18F/cf-ex-drupal8/issues/4 is closed, this
can be reverted. It's common to turn off caching locally while
doing front-end development, though, so it'd be preferable to take
the opportunity to incorporate the config_split module instead.